requirements.txt
pandas
pytest
streamlit
google-genai
pydantic>=2
pyyaml
